% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anon.R
\name{anon_calc_riesgo_global}
\alias{anon_calc_riesgo_global}
\title{Assess global disclosure risk from an \code{sdcMicro} object}
\usage{
anon_calc_riesgo_global(sdcobject)
}
\arguments{
\item{sdcobject}{An object of class \code{sdcMicro}.}
}
\value{
A data frame with the following columns:
\describe{
  \item{riesgo_global}{Unadjusted value of global risk.}
  \item{riesgo_global_chr}{Global risk expressed as a percentage
  (rounded to two decimals).}
  \item{n_esperado_reidentificaciones}{Expected number of individuals
  that could be re-identified, rounded up to the next integer.}
  \item{riesgo_global_int}{A character string providing an
  interpretative summary of the global risk.}
  \item{keyvars}{A character string with the names of the key
  variables (comma-separated).}
}
}
\description{
This function calculates global risk indicators from a given
\code{sdcMicro} object and returns them in a tidy data frame.
}
