% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anon.R
\name{anon_calc_riesgo_jerarq}
\alias{anon_calc_riesgo_jerarq}
\title{Assess hierarchical disclosure risk from an \code{sdcMicro} object}
\usage{
anon_calc_riesgo_jerarq(sdcobject)
}
\arguments{
\item{sdcobject}{An object of class \code{sdcMicro}.}
}
\value{
A data frame with the following columns:
\describe{
  \item{riesgo_jerarq}{Unadjusted value of hierarchical risk.}
  \item{riesgo_jerarq_chr}{Hierarchical risk expressed as a percentage
  (rounded to two decimals).}
  \item{n_esperado_reidentificaciones}{Expected number of individuals
  that could be re-identified, rounded up to the next integer.}
  \item{riesgo_jerarq_int}{A character string providing an
  interpretative summary of the hierarchical risk.}
  \item{keyvars}{A character string with the names of the key
  variables (comma-separated).}
}
}
\description{
This function calculates hierarchical disclosure risk indicators from a given
\code{sdcMicro} object and returns them in a tidy data frame.
}
