% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anon_internals.R
\name{anon_extraer_vars_clave}
\alias{anon_extraer_vars_clave}
\title{Extract key variables from an \code{sdcMicro} object}
\usage{
anon_extraer_vars_clave(sdcobject)
}
\arguments{
\item{sdcobject}{An object of class \code{sdcMicro}.}
}
\value{
A single character string with the names of the key variables,
separated by commas.
}
\description{
This function retrieves the names of the key variables
that were specified when creating an \code{sdcMicro} object.
}
