% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_p_tc_subpop.R
\name{calc_p_tc_subpop}
\alias{calc_p_tc_subpop}
\title{Calculate participation in care activities for a specific subpopulation of cared-for individuals}
\usage{
calc_p_tc_subpop(
  data,
  id_edit = NULL,
  cuidadora = TRUE,
  subpop = TRUE,
  actividades = NULL,
  sufijo = "_ds",
  var_edad = "c3",
  var_dep_nna = "d1",
  var_dep_adulto = "d8",
  total = FALSE,
  name_newcol = "participa"
)
}
\arguments{
\item{data}{\code{data.frame}
A dataset containing time-use information.
Each row represents an individual. Must include columns \code{interview__key} and \code{id_per}.}

\item{id_edit}{\code{numeric} or \code{character}
Value used to identify invalid responses in the dataset, which will be replaced with `NA`.
Default is \code{-777}.}

\item{cuidadora}{\code{logical} expression or \code{dplyr::filter()} condition
Filter applied to select specific subsets of caregivers from \code{data}. TRUE to skip the filter.}

\item{subpop}{\code{logical} expression or \code{dplyr::filter()} condition
Filter applied to care recipients. Only caregiving time directed to this subpopulation is considered. TRUE to skip the filter.}

\item{actividades}{\code{character} vector
Activity prefixes to include in the computation (e.g., \code{"tc1"}, \code{"tc2"}, \code{"tc3"}).}

\item{sufijo}{\code{character}
Suffix used for time variables (e.g., \code{"_ds"} for weekday data).}

\item{var_edad}{\code{character}
Name of the column containing respondents’ age.}

\item{var_dep_nna}{\code{character}
Name of the binary column identifying dependent children (PSDF).}

\item{var_dep_adulto}{\code{character}
Name of the binary column identifying dependent adults (PSDF).}

\item{total}{Logical. If TRUE, returns aggregated results.}

\item{name_newcol}{Character. Name for the output participation variable, if total = TRUE.}
}
\value{
A data frame indicating whether each caregiver provided care
  to individuals in the specified subpopulation.
}
\description{
This function identifies whether each caregiver provides care to individuals
belonging to a specified subpopulation. It can return either a detailed dataframe
(one row per caregiver–activity combination) or an aggregated version indicating
overall participation by activity.
}
\details{
When applying this function to data from Chile’s II ENUT or the Rural Women Pilot Survey,
it assumes that transport-related questions have been reprocessed to unify their prefixes.

...
}
