% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplot.R
\name{cargar_obj_tab}
\alias{cargar_obj_tab}
\title{Generate data frame with estimation data to ease the plotting process}
\usage{
cargar_obj_tab(
  mydata,
  id_estimacion_vec = "prop_cuid_inact",
  nudge_x = 0.08,
  nudge_y = 0.05,
  str_wrap = 12
)
}
\arguments{
\item{mydata}{Data frame containing complete information on all estimations.}

\item{id_estimacion_vec}{Specific ID used to identify estimations. Consists of a character
vector. Default value is \code{"prop_cuid_inact"}.}

\item{nudge_x}{Numeric vector indicating how much distance should be plotted between
the upper confidence interval and the visualization of the point estimate (Note:
this argument is relevant only for lollipop plots and has no bearing on plotting barplots).
Default value is \code{0.08}.}

\item{nudge_y}{Numeric vector indicating how much distance should be plotted between the
lower confidence interval and the visualization of the point estimate (Note:
this argument is relevant only for barplots and has no bearing on plotting lollipop plots).
Default value is \code{0.05}.}

\item{str_wrap}{Numeric vector indicating how many characters should a string possess before
automatically inserting a new paragraph. Useful for plots where categories have long
labels.}
}
\value{
A data frame containing tabular data for socio-demographic and time-use estimations.
}
\description{
This function generates a data frame with estimation data which should substantially ease the plotting process.
}
\details{
The function generates a data frame based on estimation tabulates.


The function operates in the following way:
\itemize{
  \item The function filters the original data frame for a specific \code{id_estimacion},
        so that only information on a specific estimation is preserved.
  \item The function extracts select specific columns necessary for plotting.
  \item Afterwards, generates two new columns: \code{nudge_x_upp} by adding \code{upp} and
        \code{nudge_x}, and \code{nudge_y_low} by adding \code{low} and \code{nudge_y}.
  \item The function then applies str_wrap con the val_label column.
  \item Finally, the function consolidates two specific columns containing labels
        from character type to factor type.
}
}
