% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplot.R
\name{cplot_barplot}
\alias{cplot_barplot}
\title{Plot customized ggplot2 bar plots}
\usage{
cplot_barplot(
  mydata,
  cats_group_var = "cats_group_var_label",
  media = "estimate",
  media_label = "estimate_format_asterisk",
  media_low = "low",
  media_upp = "upp",
  bar_width = 0.4,
  ci_color = "black",
  lab_x = NULL,
  lab_y = "Tiempo (horas)",
  bar_colors = c("#ccebc5", "#decbe4"),
  nudge_y_low = "nudge_y_low"
)
}
\arguments{
\item{mydata}{A data frame containing tidy data based on survey value estimations.}

\item{cats_group_var}{A vector inside the data frame denoting a grouping
variable and its categories. Default value is \code{"cats_group_var_label"}.}

\item{media}{A vector inside the data frame denoting where the function
must look for the tabulated values. In the case of the present
function, the expected value is a mean. Therefore, the default
value for this parameter is \code{"estimate"}.}

\item{media_label}{A vector inside the data frame denoting where the function
must look for labelled values. In the case of the present function, the
expected value is a labelled mean (e.g., "2,05" instead of "2.04922222").
The default value for this parameter is \code{"estimate_format_asterisk"}. The asterisk
denotes whether the estimation is reliable, weakly reliable or not reliable.}

\item{media_low}{A vector inside the data frame denoting where the function must
look for the lower tail for confidence intervals. Default value is \code{low}.}

\item{media_upp}{A vector inside the data frame denoting where the function must
look for the upper tail for confidence intervals. Default value is \code{upp}.}

\item{bar_width}{Numeric vector (ranging from 0 to 1) denoting bar
width. Default value is \code{0.4}.}

\item{ci_color}{Character vector denoting the color the confidence
intervals must have. Default value is \code{"black"}.}

\item{lab_x}{Label for the x-axis variable. Default value is \code{NULL}.}

\item{lab_y}{Label for the y-axis variable. Default value is \code{"Tiempo (horas)"},
which is only suitable for time-use estimations.}

\item{bar_colors}{Character vector denoting the color or colors for
either both or each bar. Default colors are \code{"#ccebhc5"} and \code{"#decbe4"}.}

\item{nudge_y_low}{Name of the column where the value for the distance between
point estimate display and lower confidence interval tail can be found.
Default value is \code{"nudge_y_low"}.}
}
\value{
A ggplot2 customized bar plot.
}
\description{
This function is intended to allow the generation of bar plots
with little code input as part of a workflow where tidy tabular
data is generated using \code{\link[enuttools]{cargar_obj_tab}}. If
using the expected workflow, the user does not need to specify too
many parameters, so that plots can be built with little code input.
}
\details{
This function was designed to enable the user to easily generate
ggplot2 barplots based on tidy tabular data generated by
\code{\link[enuttools]{cargar_obj_tab}} with minimal code input.
However, the function is
flexible enough so that the user can plot customized bar plots
based on tidy tabular data generated by other means, not using
\code{\link[enuttools]{cargar_obj_tab}} (the expected workflow).


The function operates in the following way:
\itemize{
  \item The function first generates a simple ggplot2 plot by
        specifying ggplot2 basics: data, mapping (x, y and fill),
        and then adding a \code{geom_bar()}.
  \item The function then adds an error bar to allow graphical
        display of confidence intervals. Note that confidence
        interval values must be already calculated and contained
        as columns inside your tidy tabular data frame.
  \item The function then adds the themes for the plot.
  \item The function then adds captions for the plot.
  \item The function then formats the appearance of the y-axis
        (composed of continuous values). The x-axis does not require
        specific formatting because its values are expected to form an ordered factor in the
        tidy data frame.
  \item Finally, the function specifies manual color categories
        for the bar colors.
}
}
\seealso{
\code{\link[ggplot2]{ggplot}},
\code{\link[enuttools]{cargar_obj_tab}}.
}
