% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplot.R
\name{cplot_lollipop}
\alias{cplot_lollipop}
\title{Plot customized ggplot2 lollipop plots}
\usage{
cplot_lollipop(
  mydata,
  prop = "estimate",
  cats_group_var_label = "cats_group_var_label",
  prop_label = "estimate_format_asterisk",
  prop_low = "low",
  prop_upp = "upp",
  lab_y = NULL,
  lab_x = "Población que participa (\%)",
  nudge_x_upp = "nudge_x_upp",
  expand = c(0, 0.02),
  limits = c(0, 1)
)
}
\arguments{
\item{mydata}{Tidy tabular data frame or tibble.}

\item{prop}{Name of the column where the point estimate for a proportion is
stored. Default value is \code{"estimate"}.}

\item{cats_group_var_label}{A vector inside the data frame denoting the (labelled)
values of a grouping variable. Default value is \code{"cats_group_var_label"}.}

\item{prop_label}{Name of the column where the point estimate for a proportion
is stored in a formatted string (e.g., a string such as \code{"25,2\%"} instead
of \code{"0.252112"}). Default value is \code{"estimate_format_asterisk"}.}

\item{prop_low}{Name of the column where the value for the lower tail of
the confidence interval for a proportion can be found. Default value
is \code{"low"}.}

\item{prop_upp}{Name of the column where the value for the upper tail of
the confidence interval for a proportion can be found. Default value
is \code{"upp"}.}

\item{lab_y}{String for the y-axis label. Default value is NULL.}

\item{lab_x}{String for the x-axis label. Default value is \code{"Población que
participa (\%)"}}

\item{nudge_x_upp}{Name of the column where the value for the distance between
point estimate display and upper confidence interval tail can be found.
Default value is \code{"nudge_x_upp"}.}

\item{expand}{Value of \code{expand} inside \code{\link[ggplot2]{scale_x_continuous}}.
Default value is \code{c(0,0.2)}.}

\item{limits}{Value of \code{limits} inside \code{\link[ggplot2]{scale_x_continuous}}.
Default value is \code{c(0,1)}.}
}
\value{
A ggplot2 customized lollipop plot.
}
\description{
This function is intended to allow the generation of lollipop plots
with variable code input as part of a workflow where tidy tabular
data is generated using \code{\link[enuttools]{cargar_obj_tab}} along with further
data wrangling and processing if necessary. If the user follows the expected workflow,
there should be no need to specify too
many parameters, so that plots themselves can be built with little code input.
But be aware that generating tidy tabular data frames can require a certain amount
of code input.
}
\details{
This function was designed to enable the user to easily generate
ggplot2 lollipop plots based on tidy tabular data generated by
\code{\link[enuttools]{cargar_obj_tab}} and further processing if necessary.


The function operates in the following way:
\itemize{
  \item The function first generates a basic mapping using the aesthetics
        \code{x = grouping} variable and \code{y = value} for the proportion, after which it
        adds two separate elements to the plot: a
        geom_segment line and a geom_point so as to build the "lollipop"
        plot image.
  \item The function then adds error bars around the point estimation
        based on values for both tails of the confidence intervals found
        on the data frame.
  \item The function then adds a text label for the point-estimation value
        located to the right of the upper confidence interval bar.
  \item The function then customizes the x-axis (x-axis is customized as
        as continuous using \code{\link[scales]{label_percent}}).
  \item The function afterwards specifies labels and themes for the plot.
}
}
\seealso{
\code{\link[ggplot2]{ggplot}},
\code{\link[enuttools]{cargar_obj_tab}}.
}
