% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplot.R
\name{cplot_lollipop_fw}
\alias{cplot_lollipop_fw}
\title{Facet-wrap customized ggplot2 lollipop plots}
\usage{
cplot_lollipop_fw(
  mydata,
  prop = "estimate",
  cats_group_var_label = "cats_group_var_label",
  prop_label = "estimate_format_asterisk",
  prop_low = "low",
  prop_upp = "upp",
  lab_y = NULL,
  lab_x = "Población que participa (\%)",
  variable = "variable",
  nudge_x_upp = "nudge_x_upp",
  expand = c(0, 0.02),
  ncol = 1,
  nrow = 4,
  limits = c(0, 1)
)
}
\arguments{
\item{mydata}{Tidy tabular data frame or tibble.}

\item{prop}{Name of the column where the point estimate for a proportion is
stored. Default value is \code{"estimate"}.}

\item{cats_group_var_label}{A vector inside the data frame denoting the (labelled)
values of a grouping variable. Default value is \code{"cats_group_var_label"}.}

\item{prop_label}{Name of the column where the point estimate for a proportion
is stored in a formatted string (e.g., a string such as \code{"25,2\%"} instead
of \code{"0.252112"}). Default value is \code{"estimate_format_asterisk"}.}

\item{prop_low}{Name of the column where the value for the lower tail of
the confidence interval for a proportion can be found. Default value
is \code{"low"}.}

\item{prop_upp}{Name of the column where the value for the upper tail of
the confidence interval for a proportion can be found. Default value
is \code{"upp"}.}

\item{lab_y}{String for the y-axis label. Default value is \code{NULL}.}

\item{lab_x}{String for the x-axis label. Default value is \code{"Población que
participa (\%)"}.}

\item{variable}{Name of the column where the variable name is stored. \code{variable}
becomes a part of the formula necessary to perform \code{\link[ggplot2]{facet_wrap}}.
Default value is \code{"variable"}.}

\item{nudge_x_upp}{Name of the column where the value for the distance between
point estimate display and upper confidence interval tail can be found.
Default value is "nudge_x_upp".}

\item{expand}{Value of \code{expand} inside \code{\link[ggplot2]{scale_x_continuous}}.
Default value is \code{c(0,0.2)}.}

\item{ncol}{Value of \code{ncol} inside \code{\link[ggplot2]{facet_wrap}}. Default
value is 1.}

\item{nrow}{Value of \code{nrow} inside \code{\link[ggplot2]{facet_wrap}}. Default
value is 4.}

\item{limits}{Value of \code{limits} inside \code{\link[ggplot2]{scale_x_continuous}}.
Default value is \code{c(0,1)}.}
}
\value{
A ggplot2 customized lollipop plot.
}
\description{
This function is intended to allow the generation of lollipop plots for multiple plots
with variable code input as part of a workflow where tidy tabular
data is generated using \code{\link[enuttools]{cargar_obj_tab}} and the basic structure
for lollipop plots is generated using \code{\link[enuttools]{cplot_lollipop}}.
If the user follows the expected workflow, there should be no need to specify too
many parameters, so that plots themselves can be built with little code input.
}
\details{
This function was designed to enable the user to easily generate
ggplot2 lollipop plots for several groups building upon the
\code{\link[enuttools]{cplot_lollipop}} function. Multiple plots are generated
by using \code{\link[ggplot2]{facet_wrap}}.


The function operates in the following way:
\itemize{
  \item The function first generates a basic lollipop plot using \code{\link[enuttools]{cplot_lollipop}}.
  \item The function then facet-wraps the plot based on the values provided by
        \code{variable}.
}
}
\seealso{
\code{\link[ggplot2]{ggplot}},
         \code{\link[enuttools]{cargar_obj_tab}},
         \code{\link[enuttools]{cplot_lollipop}}.
}
