% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_valprimer.R
\name{edit_valprimer}
\alias{edit_valprimer}
\title{Impute missing values due to validation errors in a survey dataset}
\usage{
edit_valprimer(data, data_val, error)
}
\arguments{
\item{data}{A data frame containing the original survey data.}

\item{data_val}{A data frame containing validation dummy variables. These dummies indicate
whether a specific variable in @param data has a first-level validation error.}

\item{error}{A character vector specifying the type(s) of first-level errors to impute.

Possible options are:

\itemize{
  \item \strong{error1}: Missing value. The response is missing, even though the enabling condition indicates there should be a response.
  \item \strong{error2}: Failed filter. A response is present, even though the enabling condition indicates there should be no response.
  \item \strong{error3}: Out of range. The response is outside the expected range.
}}
}
\value{
A modified version of \code{data} with the specified first-level validation values
  imputed according to their data type.
}
\description{
This function imputes observations that triggered first-level validation alerts to
missing values (-777) within a data frame. It's designed to handle various data types
(numeric, character, and date) based on a separate validation dummy data frame
}
\details{
\itemize{
  \item \strong{Numeric variables}: Imputed to \code{-777}.
  \item \strong{Character variables}: Imputed to \code{"-777"} (as a string).
  \item \strong{Date variables}: Imputed to \code{as.Date("777-01-01")}.
}
}
\section{Assumptions}{

This function relies on the following assumptions regarding the input data frames
\enumerate{
  \item \code{data_val} must be a wide data frame.
  \item \code{data_val} must contain dummy columns where \code{1} indicates a validation error for a specific variable.
  \item The variable names in \code{data_val} must follow a specific structure: error_varname,
where \code{@{param error}} is one of the error types specified (e.g., \code{"error1"}, \code{"error2"}, \code{"error3"})
and \code{varname} is the original variable name from the \code{data} data frame.
For example, if you are imputing \code{"error1"} for a variable named \code{"age"}, the dummy column
in \code{data_val} should be named \code{"error1_age"}.
}
}

