% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_calidad_internals.R
\name{eval_calidad_mean}
\alias{eval_calidad_mean}
\title{Evaluate reliability for means estimation for a given numeric variable}
\usage{
eval_calidad_mean(
  mydata,
  est_var = "t_tvaoh_oh",
  group_var = "cui_psdf_o_resp_nna",
  subpop = "TRUE",
  qst = "cut",
  id_estimacion = "mean_cuid_t_tvaoh_oh"
)
}
\arguments{
\item{mydata}{A data frame or tibble containing the survey data.}

\item{est_var}{Estimation variable. Must be quantitative, and in terms of R
it must consist of a numeric vector. Default value is "t_tvaoh_oh".}

\item{group_var}{Grouping variable. Must be categorical in nature. If no grouping
variable is required for the estimation assessment, group_var must be specified
to "NULL" (a string). Default value is "cui_psdf_o_resp_nna".}

\item{subpop}{Subpopulation variable and values for the subpopulation variable, in
string format. If no subpopulation variable is required for the estimation
assessment, subpop must be specified to "TRUE" (a string). Default value is "TRUE" (a string).}

\item{qst}{Variable denoting the questionnaire to which the estimation variable
belongs. Only possible values are "ch" or "cut". Default value is "cut".}

\item{id_estimacion}{A given string which allows identification of a specific
estimation across different settings. \code{id_estimacion} has no bearing
on the process of generating estimations and is merely intended to function
as a name. Default value is "mean_cuid_t_tvaoh_oh".}
}
\value{
A data frame containing estimation reliability information.
}
\description{
This function streamlines the process of evaluating estimation reliability for
means.
}
\details{
The function creates a data frame where an estimation variable (for a mean), a
grouping variable and a subpopulation variable are specified, along with columns
necessary for complex survey analysis (strata, primary sampling units, and weights).
The function applies the functions \code{\link[calidad]{create_mean}} and
\code{\link[calidad]{assess}} for a given estimation (numeric) variable, its
grouping variable and its subpopulation variable. It generates as its output a tibble
data frame which contains information on estimation reliability. In terms of their
reliability, estimations can be classified as reliable, weakly reliable and
non-reliable.


The function operates in the following way:
\itemize{
  \item The function generates a subpopulation variable from a given string specifying
        the subpopulation of interest;
  \item Next, the function creates a reduced data frame containing the estimation
        variable, the grouping variable (if specified), the subpopulation variable (if
        specified), and columns for strata, primary sampling units and weights;
  \item Next, the function estimates the population mean for the given variable, and
        evaluates its reliability (means can be classified into reliable, weakly
        reliable or non-reliable estimations);
  \item Finally, the function generates a data frame as its output, containing all the
        relevant information for reliability evaluation.
}
}
\seealso{
\code{\link[calidad]{create_mean}}, \code{\link[calidad]{assess}}.
}
