% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_calidad_internals.R
\name{eval_calidad_prop_completar_niveles}
\alias{eval_calidad_prop_completar_niveles}
\title{Evaluate reliability for proportions estimation for all values of a given categorical variable}
\usage{
eval_calidad_prop_completar_niveles(
  mydata,
  est_var = "p_tdnr_admnhog",
  group_var = "cui_psdf_o_resp_nna",
  subpop = "TRUE",
  qst = "cut",
  id_estimacion = "prop_cuid_p_tdnr_admnhog"
)
}
\arguments{
\item{mydata}{A data frame or tibble containing the survey data.}

\item{est_var}{Estimation variable. Must be categorical in nature, and in terms of R
it must consist in a labelled numeric vector. Default value is "p_tdnr_admnhog".}

\item{group_var}{Grouping variable. Must be categorical in nature. If no grouping
variable is required for the estimation assessment, group_var must be specified
to "NULL" (a string). Default value is "cui_psdf_o_resp_nna".}

\item{subpop}{Subpopulation variable and values for the subpopulation variable, in
string format. If no subpopulation variable is required for the estimation
assessment, subpop must be specified to "TRUE" (a string). Default value is "TRUE" (a string).}

\item{qst}{Variable denoting the questionnaire to which the estimation variable
belongs. Only possible values are "ch" or "cut". Default value is "cut".}

\item{id_estimacion}{A given string which allows identification of a specific
estimation across different settings. \code{id_estimacion} has no bearing
on the process of generating estimations and is merely intended to function
as a name. Default value is "prop_cuid_p_tdnr_admnhog".}
}
\value{
A data frame containing estimation reliability information.
}
\description{
This function streamlines the process of evaluating estimation reliability for all
values of a categorical variable (thus, works only for estimating proportions).
}
\details{
The function creates a data frame where an estimation variable (for a proportion), a
grouping variable and a subpopulation variable are specified, along with columns
necessary for complex survey analysis (strata, primary sampling units, and weights).
The function iterates the function \code{\link[enuttools]{eval_calidad_prop_nivel}} for
all values of the categorical estimation variable, thus assessing separately an
estimation for each category (in terms of reliability). In terms of their reliability,
estimations can be classified as reliable, weakly reliable and non-reliable.


The function operates in the following way:
\itemize{
  \item The function generates a data frame for all values or categories of the
        estimation variable;
  \item Next, the function creates a numeric ID for each value in the data frame;
  \item Next, the function iterates the function \code{\link[enuttools]{eval_calidad_prop_nivel}}
        for each value ID and binds the resulting list into a tibble df containing
        reliability information for all values of the categorical variable.
}
}
\seealso{
\code{\link[enuttools]{eval_calidad_prop_nivel}}.
}
