% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_calidad_internals.R
\name{eval_calidad_total_nivel}
\alias{eval_calidad_total_nivel}
\title{Evaluate reliability for totals estimation for a given variable and variable level}
\usage{
eval_calidad_total_nivel(
  mydata,
  est_var = "nivel_educ",
  group_var = "NULL",
  subpop = "cui_psdf_o_resp_nna == 1",
  qst = "ch",
  value = 1,
  id_estimacion = "tot_nivel_educ_cuid"
)
}
\arguments{
\item{mydata}{A data frame or tibble containing the survey data.}

\item{est_var}{Estimation variable. Must be categorical in nature, and in terms of R
it must consist in a labelled numeric vector. Default value is "nivel_educ".}

\item{group_var}{Grouping variable. Must be categorical in nature. If no grouping
variable is required for the estimation assessment, group_var must be specified
to "NULL" (a string). Default value is "NULL" (a string).}

\item{subpop}{Subpopulation variable and values for the subpopulation variable, in
string format. If no subpopulation variable is required for the estimation
assessment, subpop must be specified to "TRUE" (a string). Default value is
"cui_psdf_o_resp_nna == 1".}

\item{qst}{Variable denoting the questionnaire to which the estimation variable
belongs. Only possible values are "ch" or "cut". Default value is "ch".}

\item{value}{Given categorical response for the estimation variable, expressed as
a single numeric value. Default value is 1.}

\item{id_estimacion}{A given string which allows identification of a specific
estimation across different settings. \code{id_estimacion} has no bearing
on the process of generating estimations and is merely intended to function
as a name. Default value is "tot_nivel_educ_cuid".}
}
\value{
A data frame containing estimation reliability information.
}
\description{
This function streamlines the process of evaluating estimation reliability.
}
\details{
The function creates a data frame where an estimation variable (for totals estimation), a
grouping variable and a subpopulation variable are specified, along with columns
necessary for complex survey analysis (strata, primary sampling units, and weights).
A given value or category for the estimation variable is then transformed into a dummy
variable inside the data frame. The dummy variable is used to generate a totals
estimation, which is then assessed in terms of its reliability. In terms of their
reliability, estimations can be classified as reliable, weakly reliable and
non-reliable.


The function operates in the following way:
\itemize{
  \item The function generates a subpopulation variable from a given string specifying
        the subpopulation of interest;
  \item Next, the function creates a reduced data frame containing the estimation
        variable, the grouping variable (if specified), the subpopulation variable (if
        specified), and columns for strata, primary sampling units and weights;
  \item Next, the function transforms a labelled numeric vector into a factor;
  \item Next, the function transforms a given factor level into a dummy variable (with
        values 0 and 1);
  \item Next, the function estimates the population total for the given dummy, and
        evaluates its reliability (proportions can be classified into reliable, weakly
        reliable or non-reliable estimations);
  \item Finally, the function generates a data frame as its output, containing all the
        relevant information for reliability evaluation.
}
}
\seealso{
\code{\link[calidad]{create_size}}, \code{\link[calidad]{assess}}.
}
