% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suf_eval_ch.r
\name{suf_eval_ch}
\alias{suf_eval_ch}
\title{Evaluate sufficiency of information in time-use enabling questions}
\usage{
suf_eval_ch(df_valid, p_eval, level = c("person", "household"), detail = FALSE)
}
\arguments{
\item{df_valid}{A validated data frame with the characteristics described above.}

\item{p_eval}{Character vector with the variable names to check (without the
"error1_" prefix).}

\item{level}{Character string specifying the level of output:
\code{"person"} returns results at the individual level;
\code{"household"} aggregates results by \code{interview__key}.}

\item{detail}{\code{FALSE} to return only corresponding insufficiency columns.}
}
\value{
A data frame containing sufficiency indicators at the specified level.
}
\description{
The function checks, for each row, wether the set of questions required to
enable time-use modules contains sufficient data. It then aggregates this
information to the household level to assess the sufficiency of the interview.

The function should be applied to a "validated dataframe" (the output of the
\code{valprimer()} function).
}
