% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valprimer.R
\name{valprimer}
\alias{valprimer}
\title{Validate questions in a survey dataset}
\usage{
valprimer(
  datos,
  variables,
  expresiones,
  rangos,
  tipo_rangos,
  id,
  nec_justificar = FALSE,
  justificar,
  consider_edits = FALSE
)
}
\arguments{
\item{datos}{A data frame containing the survey data.}

\item{variables}{A character vector with the names of the variables to validate.}

\item{expresiones}{A character vector of R expressions specifying the enabling
conditions for each variable.}

\item{rangos}{A character vector indicating the allowed values or range for
each variable. For text-type variables, this should indicate the minimum number
of characters expected.}

\item{tipo_rangos}{A character vector indicating the type of range for each variable.
Possible values: \code{normal}, \code{texto}, \code{tiempo} or \code{tiempo0}.}

\item{id}{The name of the column containing the unique identifier for each case.}

\item{nec_justificar}{Logical value. If \code{TRUE} indicate the necessity of justify some variables.}

\item{justificar}{A character vector with the name of the validation columns to justify.
For example: \code{c("error1_se3_2", "error1_tc1_n2_ds")}. This imputes 0 to the columns.}

\item{consider_edits}{Logical value. If \code{TRUE} a \code{-777} (missing value code) is inserted in the variable range.}
}
\value{
A wide \code{data.frame} with dummy columns indicating validation errors.
}
\description{
This function creates a data frame with dummy columns indicating whether each
variable in a survey dataset meets first-level validation criteria.
}
\details{
The first-level validation criteria are:

\enumerate{
  \item The cell value is allowed by the enabling condition associated with the column.
  \item The cell value falls within the expected range for the column.
}

These criteria can produce the following types of errors:
\itemize{
  \item \strong{error1}: Missing value. The response is missing, even though the enabling condition indicates there should be a response.
  \item \strong{error2}: Failed filter. A response is present, even though the enabling condition indicates there should be no response.
  \item \strong{error3}: Out of range. The response is outside the expected range.
}

For each variable evaluated, the function generates up to three new columns
(one per error type). If \eqn{c} columns are validated, the output will contain
\eqn{c * 3} columns. Each column is a dummy variable indicating with \eqn{1} whether
the error occurred for that case.

The function assumes the existence of a structured metadata input that includes
the question name (column name), enabling condition, valid range, and type of range.
}
