% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valprimer_check.R
\name{valprimer_check}
\alias{valprimer_check}
\title{Check Activated First-Level Validation}
\usage{
valprimer_check(
  data,
  data_val,
  vars,
  exps,
  vars_exp,
  ranges,
  error,
  var_constant,
  var_to_check,
  id
)
}
\arguments{
\item{data}{\code{data.frame}. Original dataset containing the variables of interest.}

\item{data_val}{\code{data.frame}. Dataset containing validation variables with
indicators (0/1), typically the output from the \code{valprimer()} function.}

\item{vars}{\code{character}. Vector with all survey variables.}

\item{exps}{\code{character}. Vector with all enabling conditions of the survey in R format.}

\item{vars_exp}{\code{character}. Vector with the variables required for
each enabling condition (comma-separated).}

\item{ranges}{\code{character}. Vector with the valid ranges of all survey variables.}

\item{error}{\code{character}. A string specifying the type of error to review.
Can be \code{"error1"}, \code{"error2"}, or \code{"error3"}.}

\item{var_constant}{\code{character}. Vector with constant variables to be included
in the review (e.g., sex, age).}

\item{var_to_check}{\code{character}. Name of the variable to review.}

\item{id}{\code{character}. Name if the unique survey identifier.}
}
\value{
A \code{data.frame} with the following elements:
\itemize{
  \item \code{var}: Variable being checked.
  \item \code{expr}: Enabling condition for the variable.
  \item \code{range}: Valid range of the variable.
  \item Characterization variables, the variables used in the condition,
        and the variable itself.
  \item The survey identifier for cases that triggered the validation.
}
}
\description{
Creates a \code{data.frame} with the information required to review
cases that triggered a first-level validation in the survey data.
}
\details{
Only cases that activated the selected validation (value equal to 1
in \code{data_val}) are included in the output.

The function assumes the existence of a metadata source containing
information about all variables, enabling conditions, variables used in
the conditions, and survey ranges.
}
\examples{
library(dplyr)

# Example of metadata
metadata <- tibble(
  vars = c("age", "working", "sex"),
  exps = c(TRUE, "age >= 18", "TRUE"), # all participants in the survey must respond age and sex
  vars_exp = c(NA, "age", NA),
  ranges = c("c(0:110)", "c(0:1)", "c(1:2)")
)

# Example survey data
data <- tibble(
  id = 1:5,
  age = c(18, 46, 12, 22, 16),
  working = c(NA, 1, 0, 1, 0),
  sex = c("M", "F", "M", "F", "M")
)

# Example validation output (from valprimer)
data_val <- tibble(
  id = 1:5,
  error1_age = c(0, 0, 0, 0, 0),
  error1_working = c(1, 0, 0, 0, 0), # Example of an activation
  error1_sex = c(0, 0, 0, 0, 0),
  error2_age = c(0, 0, 0, 0, 0),
  error2_working = c(0, 0, 0, 0, 0),
  error2_sex = c(0, 0, 0, 0, 0),
  error3_age = c(0, 0, 0, 0, 0),
  error3_working = c(0, 0, 0, 0, 0),
  error3_sex = c(0, 0, 0, 0, 0)
)

# Run check for error1 on variable 'working'
valprimer_check(
  data = data,
  data_val = data_val,
  vars = metadata$vars,
  exps = metadata$exps,
  vars_exp = metadata$vars_exp,
  ranges = metadata$ranges,
  error = "error1",
  var_constant = "sex",
  var_to_check = "working",
  id = "id"
)

}
