% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valseg.R
\name{valseg}
\alias{valseg}
\title{Validate questions in a survey dataset}
\usage{
valseg(data, name, expr, id, nec_justificar = FALSE, justificar)
}
\arguments{
\item{data}{A data frame containing the survey data.}

\item{name}{A character vector with the names identifying the validations to be
executed.}

\item{expr}{A character vector of R expressions specifying the conditions for each
validation.}

\item{id}{The name of the survey dataset column containing the unique identifier
variable for each row.}

\item{nec_justificar}{Logical value. If TRUE indicates necessity of justify some columns}

\item{justificar}{A character vector with columns to justify (imputes 0 to the result).
Example: c("inc_rph_r8", "inc_ch_r71")}
}
\value{
A wide \code{data.frame} with dummy columns indicating validation errors. Each
column contains a specific validation.
}
\description{
This function creates a data frame with dummy columns indicating whether each
second-level validation condition for a survey dataset is met.
}
\details{
For each validation condition evaluated, the function generates a new column.
If \eqn{c} conditions are evaluated, the output will contain \eqn{c + 1} columns, where
the additional column is an id variable. Each column is a dummy variable
that indicates with \eqn{1} when an observation does not meet the validation criteria.

The function assumes the existence of a structured metadata input that includes
the validation names and validation conditions.
}
